requirements.txt
streamlit==1.31.0
pandas==2.0.3
numpy==1.24.3
plotly==5.17.0
openai==1.3.7
sqlalchemy==2.0.23
sqlite3
openpyxl==3.1.2
python-dotenv==1.0.0
streamlit-chat==0.1.1
streamlit-aggrid==0.3.4
psycopg2-binary==2.9.9
pymysql==1.1.0
